move annotate, fix azure package name
