+++ cmd/bosun/expr/azure/azureai.go
@@ -1,4 +1,4 @@
-package expr
+package azure
 
 import (
 	"context"
